requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
python-dotenv==1.1.1

# commands 
# uvicorn app.main:app --reload


